+++ nl/lib/nl/error.rb
@@ -2,4 +2,9 @@ module Nl
   class Error < StandardError; end
   class ProtocolViolation < Error; end
   class ExhaustedSequenceNumber < Error; end
+  class TimeoutError < Error; end
+  class ClosedError < Error; end
+  class NotificationLossError < Error; end
+  class UnknownMulticastGroupError < Error; end
+  class UnresolvedMulticastGroupError < Error; end
 end